Update setting-pat-ubuntu.md
Added Link To Alexey Pertsev's case, an Open-Source Dev who wrote the Crypto Mixer Tornado-Cash. Governments are prosecuting a Developer who has written a privacy-focused software that was used by Lazarus and was not linked to Lazarus. Since they cannot focus on making laws that'll protect the crypto-markets, they decided to go on a power trip and attack innocent people because they're incompetent and cannot defend their people against a small and poor country so they they decided to turn a Developer into their Scapegoat and then flaunt his head to the people in hopes that they'll look strong. Pathetic.